\name{lfstat-package}
\alias{lfstat-package}
\alias{lfstat}
\docType{package}
\title{
Calculates Low Flow Statistics for daily stream flow data
}
\description{
The "Manual on Low-flow Estimation and Prediction", published by the World Meteorological Organisation (WMO), gives an idea how to analyse stream flow data focusing on low-flow issues. This packages gives functions to compute the descriped statistics and produce plots similar to the one in the manual.
}
\details{
\tabular{ll}{
Package: \tab lfstat\cr
Type: \tab Package\cr
Version: \tab 0.1\cr
Date: \tab 2011-08-24\cr
License: \tab GPL version 2 or later\cr
LazyLoad: \tab yes\cr
}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
Daniel Koffler <daniel.koffler@boku.ac.at>
}
\references{
Manual on Low-flow Estimation and Prediction, Operational Hydrology Report No. 50, Koblenz 2009
}

\keyword{low flow, WMO}
\seealso{
}
\examples{
}
