\name{ngaruroro}
\docType{data}
\alias{ngaruroro}
\title{Daily stream flow data from Ngaruroro River in New Zealand}
     \description{
    This data set gives the streamflow records from 20/09/1963 to
    31/12/2000. The data structure is a low flow object (lfobj) as used
    in the package lfstat.
     }
     \usage{data(ngaruroro)}
     \format{A lfobj \code{\link{createlfobj}}}
     \source{Manual on Low-flow Estimation and Prediction, Operational Hydrology Report No. 50, Koblenz 2009}
     \references{
       Manual on Low-flow Estimation and Prediction, Operational Hydrology Report No. 50, Koblenz 2009
     }
     \keyword{datasets}