\name{trace_value}
\alias{trace_value}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{Draw Paths to Points perpendicular to Coordinate Axis}
\description{
To depict the distances in \code{x} and \code{y} direction to a point, draw lines and labels.}
\usage{
trace_value(x, y, digits = 0, lab.x = x, lab.y = y, prefix = "",
            suffix = "", cex = 0.75, col = "blue", lty = 2, ...)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{x}{numeric vector of x coordinates}
  \item{y}{numeric vector of y coordinates}
  \item{digits}{vector of length one or two, giving the number of digits used for rounding the label of the \code{x} and \code{y} coordinate.}
  \item{lab.x}{character vector of length one. Label of the \code{x} coordinate.}
  \item{lab.y}{character vector of length one. Label of the \code{y} coordinate.}
  \item{prefix}{vector of length one or two, text printed before the label of the \code{x} and \code{y} coordinate.}
  \item{suffix}{vector of length one or two, text printed after the label of the \code{x} and \code{y} coordinate.}
  \item{cex}{character expansion factor}
  \item{col}{colour used for text and lines}
  \item{lty}{linetype}
  \item{\dots}{other graphical parameters, passed on to \code{lines}, \code{points} and \code{text}.}
}
\details{This function is vectorised over \code{x} and \code{y}.}
\value{This function is used for its side effect.}
\examples{
x <- c(-2, 3)
curve(sin, -2*pi, 2*pi, xname = "t")
trace_value(x, sin(x), digits = c(0, 1))
}
