\name{ngaruroro}
\docType{data}
\alias{ngaruroro}
\title{Daily stream flow data from Ngaruroro River in New Zealand}
     \description{
    This data set gives the streamflow records from 20/09/1963 to
    31/12/2000. The data structure is a low flow object (lfobj) as used
    in the package lfstat. The rivers hydrological year starts with December.
     }
     \usage{data(ngaruroro)}
     \format{A lfobj \code{\link{createlfobj}}}
     \source{Gustard, A. & Demuth, S. (2009) (Eds) Manual on Low-flow Estimation and Prediction. Operational Hydrology Report No. 50, WMO-No. 1029, 136p.}
     \references{
       Gustard, A. & Demuth, S. (2009) (Eds) Manual on Low-flow Estimation and Prediction. Operational Hydrology Report No. 50, WMO-No. 1029, 136p.
     }
     \keyword{datasets}